refactor(home content): receive each data as param, callbacks can stay as class
